+++ shaders/utils/getAmbientLight.glsl
@@ -37,7 +37,8 @@ vec3 getAmbientLight(float ambientBrightness  ARGS_OUT) {
 		+ ambientSunsetLight * lightMult;
 	
 	#include "/import/screenBrightness.glsl"
-	ambientLight = mix(CAVE_AMBIENT_COLOR * (0.8 + 0.4 * screenBrightness), ambientLight, ambientBrightness);
+	ambientBrightness *= 0.9 + 0.2 * screenBrightness;
+	ambientLight = mix(CAVE_AMBIENT_COLOR * (0.6 + 0.8 * screenBrightness), ambientLight, ambientBrightness);
 	#ifdef NETHER
 		ambientLight *= vec3(1.0, 0.5, 0.3);
 	#endif